Increase `open:` query performance on large repositories (#88)
With the `tensorflow` repository, this takes `open:` query time down
from ~1 second to ~20 milliseconds.

The central idea behind this change is introducing dummy documents in
the file index, representing directories. This allows us to search for
folders, without having to infer their existence due to other documents.

Additionally, a bug with open queries of the form
`some/valid/file/with/trailing/slash.txt/` resulting in panics was fixed.