Prevent form submission during Japanese input composition (#270)
* Prevent form submission during Japanese input composition

Made improvements to the handling of Japanese input by using the onCompositionStart and onCompositionEnd events to prevent form submission during the composition phase. Specifically, I added a composing state and modified the handleEnter function to check this state before allowing form submission. This ensures that the form is not submitted until the user has finished composing their input.

* Fix a lint issue