Implement tool-based answer module
Here, we utilize the LLM to answer questions by giving it several
actions to select from. Each action operates on our data in some way, for
example the `path` action will search for files. The responses to these
actions build up knowledge for the LLM such that it can answer questions
better.

The `answer` module was entirely rewritten, and split up. A new
`llm_gateway` module holds an interface to the "Answer API", and we
also split prompts out into their own module.