fix semantic code and path search filters (#470)
the repo filter on these two types of search accept a human-readable
notion of `RepoRef`. for a `RepoRef` of the form local//home/np/bloop, a
valid repo filter would be:

    repo:bloop

whereas, the generated filter was:

    repo:local//home/np/bloop

which resulted in 0 search results for these two tools. the former is
produced by `RepoRef::display_name`.