Fix tokenization on large unicode characters (#574)
We now ensure that tokenization does not result in malformed UTF-8. This
is tricky to achieve with the `tiktoken-rs` API as errors in the
iterator APIs do not report the size of invalid tokens. Instead we
resort to tokenizing whole strings and using a back-off strategy to get
the largest valid string for a truncated token list.