Yield before final `step` return
Calls to `self.update` with a conclusion are generally the final `await`
point in the last `step` call created during an `/answer` request. These
final `await` points typically do not return `Pending` as the channel
utilizes an internal queue, so `step` concludes early, breaking out of the
driving loop before the final update ever gets a chance to be processed.

To solve this, we `yield_now().await` right before ending the final
`step` call, so that any backed up messages in the queue are given a
chance to be processed.