Empty pending exchange queue after each step
If a step internally pushes an exchange to the exchange channel while
executing, there is a chance that this exchange will be missed in the
`Agent` driving loop. This can happen if the update does not back up the
channel, leading to an immediate success that returns `Poll::Ready`, and
if there are no further await points in the final `step` call which
return `Poll::Pending`.

We fix this by emptying out the exchange channel after every step,
picking up all stream items that immediately resolve.