Replace base64 encoding with markdown encoding
Now, rather than encoding to base64 when we send to the front-end, we
encode to markdown instead. This has the benefit that the model can
still generate XML, but the front-end only sees simple markdown.

The downside here is that our markdown blocks are stored on the
exchange, which adds to the LLM history as successive queries are made.
This seems to influence the model and encourage it to use markdown in
*new* queries, despite this being invalid.