Expand code chunks to try and maximize context (#736)
We now try to expand code chunks to fill 50% of the context, using a greedy iterative algorithm:

1. Calculate per-chunk expansion: we target a small line increment (100 lines as currently implemented), and divide this number to obtain per-chunk line increments.
2. We expand each chunk both up and down, according to this increment
3. We merge overlapping chunks
4. Repeat until we reach our 50% threshold

This iterative approach avoids problems with overlapping spans, and is fairly simple to follow.

---

* Remove special cased single-file expansion

Previously, we used this as a heuristic to try and fill the context when
there is only one file. Now that we have proper expansion, we can drop
this heuristic expansion, and have better quality context that always
contain the original code chunks.

* Use selected path aliases in `none`

This also lowers the chunk to prompt ratio slightly to avoid generating
`system` messages that are too long.

* Revert alias filtering in context generation

This doesn't work well, so we revert to the old behaviour. While the old
behaviour is less precise, we rely on it to show chunks which are not
directly selected by the model. This is part of a bigger issue and
should be investigated separately.

* Trim chunks after canonicalization

* Deduplicate code chunks even when context is full

* Increase answer history headroom