Add `/studio/import`
This route accepts a query parameter in the form `?thread_id=...`. It
must be called from a logged-in state. When successful, it will return a
string for the new studio ID.

A new studio is created with a name equal to the the conversation title,
and its messages and context are imported, collapsing exchange code
chunks and extracting query/answer pairs.

Additionally, `studio.context` JSON objects have been modified such that
`branch` is now optional.