Count message and prompt tokens in studio (#900)
Returned token count objects now look like:

```
{ total: number, messages: number, per_file: [number] }
```

Note the addition of a new `messages` field. `messages` contains the token
count of the current conversation, and `total` now also includes
`messages` in the sum.

Additionally, we count the number of tokens that are part of the base
prompt, with the use of an empty context, and include this in the new
`messages` count, and thus `total` count.