Track studio generation and import events (#916)
**WARNING: This will wipe the `studios` table when run. Studios are
still an in-dev feature so this should not have consequences for any
real users.**

---

Now, we track successful studio import calls, and generation requests
alongside associated response strings.

A few things were changed here:

- A new `StudioEvent` type was introduced, following the existing
  pattern for `QueryEvent`
- Event filtering was only used with a boolean telemetry check, so this
  was cleaned up in order to support telemetry checks for non-query
  events
- The `sqlx/uuid` feature was enabled so that we can natively use UUIDs
  in queries. **This required a migration to move away from `TEXT`-based
  UUIDs to `BLOB`-based UUIDs. All entries in `studios` are dropped.**

Note that `QueryEvent` and `StudioEvent` are intentionally not merged,
as the `QueryEvent` type must remain stable because our analytics
database schema relies on the type in order to describe table columns.

Rather, a similar event was added but specialized to Studio-related
data. In the long term, we may want to consider a refactor to merge
these two types, but this may require an analytics migration.