Add proxy billing routes (#933)
Here, we add 2 new routes:

- `GET /api/quota`: This will return an object like
  `{ upgraded: boolean, used: number, allowed: number }`
- `GET /api/quota/create-checkout-session`: This will create a checkout
  session if the user is new, or if they have already purchased a plan,
  return a link to adjust the plan. The returned object looks like
  `{ url: string }`, with a Stripe URL.

Additionally, each call to `/studio/:id/generate` will be quota-limited,
such that after `used` hits the `allowed` request count, future calls will
fail. The front-end can ensure we don't hit an error here by disabling the
`Generate` button when the `used` value is above `allowed`.

Note that the Answer API must be quota-enabled for this to work.
Instructions are listed in commit 4902af68 for the relevant repo.