Limit maximum number of steps, not exchanges (#1073)
Currently, the agent will avoid building up the context further after
reaching a maximum number of *exchanges*, not steps. In practice, this
means that after the 10th exchange, future exchanges will always be
answered instantly based on existing context.

This fixes the logic so that we instead avoid creating more than 10
search steps in a single exchange.